requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
